+++ models/index.js
@@ -2,10 +2,12 @@ const AccountModel = require('./account');
 const ProviderModel = require('./provider');
 const ProductModel = require('./product');
 const DeliveryOrderModel = require('./deliveryorder');
+const PriceModel = require('./price');
 
 module.exports = {
   AccountModel,
   ProviderModel,
   ProductModel,
   DeliveryOrderModel,
+  PriceModel,
 };