+++ test/models/provider.test.js
@@ -6,7 +6,25 @@ const models = require('../..');
 
 const fakeDatabase = require('../test-db')(models.mongoose);
 
-const {ProviderModel} = models;
+const { ProviderModel } = models;
+
+/**
+ * Test collection
+ * @param {Object} document
+ * @param {Object} mock
+ * @private
+ */
+const _checkCreated = (document, mock) => {
+  expect(document._id.toString()).toBe(mock._id);
+  expect(document.address).toBe(mock.address);
+  expect(document.email).toBe(mock.email);
+  expect(document.name).toBe(mock.name);
+  expect(document.phone).toBe(mock.phone);
+  expect(document.businessName).toBe(mock.businessName);
+  expect(document.cif).toBe(mock.cif);
+  expect(document.hasRate).toBe(mock.hasRate);
+};
+
 
 describe('provider', () => {
   beforeAll(() => fakeDatabase.connect());
@@ -26,13 +44,7 @@ describe('provider', () => {
     test('It should contain all the properties specified in the model', async () => {
       const document = await ProviderModel.findOne();
 
-      expect(document._id.toString()).toBe(providerData._id);
-      expect(document.address).toBe(providerData.address);
-      expect(document.email).toBe(providerData.email);
-      expect(document.name).toBe(providerData.name);
-      expect(document.phone).toBe(providerData.phone);
-      expect(document.businessName).toBe(providerData.businessName);
-      expect(document.cif).toBe(providerData.cif);
+      _checkCreated(document, providerData);
     });
   });
 
@@ -52,19 +64,8 @@ describe('provider', () => {
     test('Check accounts created', async () => {
       const documentList = await ProviderModel.find({});
 
-      expect(documentList[0].address).toBe(twoProviders.providers[0].address);
-      expect(documentList[0].email).toBe(twoProviders.providers[0].email);
-      expect(documentList[0].name).toBe(twoProviders.providers[0].name);
-      expect(documentList[0].phone).toBe(twoProviders.providers[0].phone);
-      expect(documentList[0].businessName).toBe(twoProviders.providers[0].businessName);
-      expect(documentList[0].cif).toBe(twoProviders.providers[0].cif);
-
-      expect(documentList[1].address).toBe(twoProviders.providers[1].address);
-      expect(documentList[1].email).toBe(twoProviders.providers[1].email);
-      expect(documentList[1].name).toBe(twoProviders.providers[1].name);
-      expect(documentList[1].phone).toBe(twoProviders.providers[1].phone);
-      expect(documentList[1].businessName).toBe(twoProviders.providers[1].businessName);
-      expect(documentList[1].cif).toBe(twoProviders.providers[1].cif);
+      _checkCreated(documentList[0], twoProviders.providers[0]);
+      _checkCreated(documentList[1], twoProviders.providers[1]);
     });
   });